.claude/skills/fastapi-backend/assets/boilerplate/pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "FastAPI backend with uv"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "httpx>=0.28.1",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
